refacto: skills spi
